create the menu controller for the client

+++ database/migrations/2020_03_14_124901_create_meals_table.php
@@ -16,7 +16,7 @@ class CreateMealsTable extends Migration
         Schema::create('meals', function (Blueprint $table) {
             $table->bigIncrements('id');
             $table->string('title');
-            $table->double('price');
+            $table->bigInteger('price');
 //            $table->time('time');
             $table->text('describe');
             $table->text('image');
